Code gen for internally tagged union variant converters
