+++ tests/Tests/_VerifySnapshots/IndexSettingsSerializationTests.CanSerializerIndexSettingsWithCustomAnalyzer.verified.txt
@@ -1,0 +1,13 @@
+{
+  analysis: {
+    analyzer: {
+      whitespace_lowercase: {
+        filter: [
+          lowercase
+        ],
+        tokenizer: whitespace,
+        type: custom
+      }
+    }
+  }
+}
